correct integration

+++ scripts/footprint_by_cluster.sh
@@ -29,7 +29,6 @@ if [ ! -z "$HINT_PATH" || ! -d "$HINT_PATH" ]; then
     fi
 fi
 
-#input_peak=${OUTPUT_DIR}/peaks/${PEAK_CALLER}/${OUTPUT_PREFIX}.${MAPPING_METHOD}_features_BlacklistRemoved.bed  ## suppose path for peak file
 
 ## select regions (peaks) by DA ##
 cluster1=${bam1/.bam/}
@@ -38,15 +37,17 @@ cluster2=${bam2/.bam/}
 cluster2=$(basename $cluster2)
 seurat_obj=${OUTPUT_DIR}/downstream_analysis/${CELL_CALLER}/seurat_obj.rds
 
-mkdir -p TMP0
 cl1=${cluster1/cluster_/}
 cl2=${cluster2/cluster_/}
 
-echo "do DA..."
-${R_PATH}/Rscript --vanilla ${curr_dir}/src/runDA.R $seurat_obj TMP0 $cl1 $cl2 LR 
-echo "DA done succefully!"
 
-#echo 'library(data.table); dd = fread("TMP0/differential_peak_cluster_table.txt");
+input_peak=${OUTPUT_DIR}/downstream_analysis/${CELL_CALLER}/differential_peak_cluster_table.txt
+if [ ! -e "$input_peak" ]; then
+    echo "do DA..."
+    mkdir -p TMP0
+    ${R_PATH}/Rscript --vanilla ${curr_dir}/src/runDA.R $seurat_obj TMP0 all others LR 
+    echo "DA done succefully!"
+#    echo 'library(data.table); dd = fread("TMP0/differential_peak_cluster_table.txt");
 #                dd[, "chr" := unlist(strsplit(peak, "-"))[1], by = peak];
 #                dd[, "start" := as.integer(unlist(strsplit(peak, "-"))[2]), by = peak];
 #                dd[, "end" := as.integer(unlist(strsplit(peak, "-"))[3]), by = peak];
@@ -54,12 +55,10 @@ echo "DA done succefully!"
 #                setkey(dd, chr, start)
 #                write.table(dd, file = "TMP0/select_peaks.bed", row.names = F, col.names = F,
 #                            quote = F, sep = "\t") 
-#' > TMP0/tmp.R
-#${R_PATH}/R --vanilla < TMP0/tmp.R
-
-
-#input_peak=TMP0/select_peaks.bed
-input_peak=output/downstream_analysis/FILTER/differential_peak_cluster_table.txt
+#    ' > TMP0/tmp.R
+#    ${R_PATH}/R --vanilla < TMP0/tmp.R
+    input_peak=TMP0/differential_peak_cluster_table.txt
+fi
 
 
 unset PYTHONPATH
@@ -71,6 +70,8 @@ ${HINT_PATH}/rgt-hint footprinting --atac-seq --paired-end --organism=${GENOME_N
 ${HINT_PATH}/rgt-hint footprinting --atac-seq --paired-end --organism=${GENOME_NAME} --output-location=$output_dir  \
         --output-prefix=${cluster2} $bam2 $input_peak &
 
+wait
+
 echo "overlap with motif annotation ... "
 ${HINT_PATH}/rgt-motifanalysis matching --organism=${GENOME_NAME} --input-files ${output_dir}/${cluster1}.bed ${output_dir}/${cluster2}.bed --output-location $output_dir
 
@@ -78,7 +79,10 @@ echo "Differential binding analysis ..."
 ${HINT_PATH}/rgt-hint differential --organism=${GENOME_NAME} --bc --nc 4 --mpbs-file1=${output_dir}/${cluster1}_mpbs.bed \
     --mpbs-file2=${output_dir}/${cluster1}_mpbs.bed --reads-file1=$bam1 --reads-file2=$bam2 --condition1=$cluster1 --condition2=$cluster2 --output-location=${output_dir}/${cluster1}_${cluster2}
 
-rm -r TMP0
+if [ -d "$TMP0" ]; then
+    rm -r TMP0
+fi
+
 echo "Footprinting Analysis Done!"
